backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Website cloning service using LLM"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "anthropic>=0.18.1",
    "python-dotenv>=1.0.0",
]
